improve resource deletion (#2244)

+++ controllers/argocd/applicationset.go
@@ -541,6 +541,9 @@ func (r *ReconcileArgoCD) reconcileApplicationSetClusterRole(cr *argoproj.ArgoCD
 
 	// ArgoCD not cluster scoped, cleanup any existing resource and exit
 	if !allowed {
+		if !argoutil.CheckClusterRoleOwnership(existingClusterRole, cr) {
+			return existingClusterRole, nil
+		}
 		argoutil.LogResourceDeletion(log, existingClusterRole, "argocd not cluster scoped")
 		err := r.Delete(context.TODO(), existingClusterRole)
 		if err != nil {
@@ -605,6 +608,9 @@ func (r *ReconcileArgoCD) reconcileApplicationSetClusterRoleBinding(cr *argoproj
 
 	// ArgoCD not cluster scoped, cleanup any existing resource and exit
 	if !allowed {
+		if !argoutil.CheckClusterRoleBindingOwnership(existingClusterRB, cr) {
+			return nil
+		}
 		argoutil.LogResourceDeletion(log, existingClusterRB, "argocd not cluster scoped")
 		err := r.Delete(context.TODO(), existingClusterRB)
 		if err != nil {
